fix: 1) list_source_filings 异常吞噬过宽 2) Streamlit 后台线程与 st.session_state 的并发安全没有被测试覆盖; 6)filing_tab.py 体量过大，接近 god module ; 10) 更新 dayu/fins/README.md
